Move off buildpack-deps and use ruby-build to install ruby
This commit makes some changes in response to some discovered when we used Clair to scan this image.

This commit uses `ubuntu:trusty` as the parent image instead of `buildpack-deps:trusty`.
The disadvantage of this is that we have to install some simple dependencies like `curl`, `libssl`, etc.
The advantage is we don't pick up a lot of tools that we don't use that have a lot of CVEs (example ImageMagick`).

This commit also uses `ruby-build` to install ruby instead of installing from the source.
Before this commit, Clair was able to find a significant number of vulnerabilities in the tools we used to compile ruby.
Using ruby-build enables us to avoid that problem by only including specifically what is needed to install ruby.