Remove BUNDLE_APP_CONFIG env variable
Prior to 04b90fb1f2bcd915b3bc5c40af7624a23e63736a, the bundler app
config was set to the GEM_HOME. This caused problems after that was
removed in favor of letting rvm manage the `GEM_HOME` variable because
the result was `BUNDLER_APP_CONFIG` being set to an empty string. As a
result of this, bundler was looking for `./config` when trying to load
it's configs. If an app had a `config` file or directory, it would cause
any `bundle` commands to fail.